Refactor URL utilities to UrlHelper and update references
Moved URL-related utility methods from Helper to a new UrlHelper class in Moonglade.Utils. Updated all usages across the codebase to reference UrlHelper instead of Helper for route link generation, URL extraction, canonical URL resolution, and related utilities. This improves code organization and separation of concerns.
