+++ src/Moonglade.Core/PostFeature/CreatePostCommand.cs
@@ -59,7 +59,7 @@ public class CreatePostCommandHandler(
             IsOutdated = request.Payload.IsOutdated,
         };
 
-        post.RouteLink = Helper.GenerateRouteLink(post.PubDateUtc.GetValueOrDefault(), request.Payload.Slug);
+        post.RouteLink = UrlHelper.GenerateRouteLink(post.PubDateUtc.GetValueOrDefault(), request.Payload.Slug);
 
         await CheckSlugConflict(post, ct);
