+++ src/Moonglade.Core/PostFeature/UpdatePostCommand.cs
@@ -124,7 +124,7 @@ public class UpdatePostCommandHandler(
         post.IsFeatured = postEditModel.Featured;
         post.HeroImageUrl = string.IsNullOrWhiteSpace(postEditModel.HeroImageUrl) ? null : SecurityHelper.SterilizeLink(postEditModel.HeroImageUrl);
         post.IsOutdated = postEditModel.IsOutdated;
-        post.RouteLink = Helper.GenerateRouteLink(post.PubDateUtc.GetValueOrDefault(), postEditModel.Slug);
+        post.RouteLink = UrlHelper.GenerateRouteLink(post.PubDateUtc.GetValueOrDefault(), postEditModel.Slug);
     }
 
     private static void UpdateCats(PostEntity post, Guid[] catIds, CancellationToken ct)
