+++ src/Moonglade.Web/Pages/Post.cshtml
@@ -40,7 +40,7 @@
 
     @if (BlogConfig.AdvancedSettings.EnableWebmention)
     {
-        <link rel="webmention" href="@(Helper.ResolveCanonicalUrl(BlogConfig.GeneralSettings.CanonicalPrefix, "webmention"))">
+        <link rel="webmention" href="@(UrlHelper.ResolveCanonicalUrl(BlogConfig.GeneralSettings.CanonicalPrefix, "webmention"))">
     }
 }
 
@@ -60,7 +60,7 @@
     <meta property="og:description" content="@Model.Post.ContentAbstract" />
     @if (!string.IsNullOrWhiteSpace(BlogConfig.GeneralSettings.CanonicalPrefix))
     {
-        <meta property="og:url" content="@(Helper.ResolveCanonicalUrl(BlogConfig.GeneralSettings.CanonicalPrefix, HttpContext.Request.Path))" />
+        <meta property="og:url" content="@(UrlHelper.ResolveCanonicalUrl(BlogConfig.GeneralSettings.CanonicalPrefix, HttpContext.Request.Path))" />
     }
 }
